requirements.txt
pandas
numpy
matplotlib
psycopg2-binary
sqlalchemy
seaborn
python-dotenv
